+++ elenasrc2/elenasm/inlineparser.h
@@ -1,7 +1,7 @@
 //---------------------------------------------------------------------------
 //		E L E N A   P r o j e c t:  ELENA VM Script Engine
 //
-//                                             (C)2011-2019, by Alexei Rakov
+//                                             (C)2011-2020, by Alexei Rakov
 //---------------------------------------------------------------------------
 
 #ifndef inlineparserH
@@ -12,26 +12,38 @@
 namespace _ELENA_
 {
 
-// --- InlineScriptParser ---
+// --- VMTapeParser ---
 
 class VMTapeParser : public _Parser
 {
-   //MessageMap _verbs;
+//   //MessageMap _verbs;
    TempString _postfix;
+//
+//   //int mapVerb(ident_t literal);
+//
+//   void writeSubject(TapeWriter& writer, ident_t message);
+//   bool writeObject(TapeWriter& writer, int state, ident_t token);
+//   bool writeArgument(TapeWriter& writer, int state, ident_t token);
+//   bool writeArray(TapeWriter& writer, int state, ident_t token);
+//   bool parseMessage(ident_t message, IdentifierString& reference);
+//   bool writeMessage(TapeWriter& writer, ident_t message, int command);
+//   bool writeExtension(TapeWriter& writer, ident_t message, int command);
+   //void skipArgumentList(_ScriptReader& reader);
+   //void writeArgumentList(_ScriptReader& reader, ScriptBookmark& bm, TapeWriter& writer);
 
-   //int mapVerb(ident_t literal);
+   int writeBuildScriptArgumentList(_ScriptReader& reader, ScriptBookmark terminator, Stack<ScriptBookmark>& callStack, 
+                                    TapeWriter& writer);
+   void writeBuildScriptStatement(_ScriptReader& reader, ScriptBookmark terminator, Stack<ScriptBookmark>& callStack, 
+                                    TapeWriter& writer);
 
-   void writeSubject(TapeWriter& writer, ident_t message);
-   bool writeObject(TapeWriter& writer, int state, ident_t token);
-   bool writeArgument(TapeWriter& writer, int state, ident_t token);
-   bool writeArray(TapeWriter& writer, int state, ident_t token);
-   bool parseMessage(ident_t message, IdentifierString& reference);
-   bool writeMessage(TapeWriter& writer, ident_t message, int command);
-   bool writeExtension(TapeWriter& writer, ident_t message, int command);
-   void parseStatement(_ScriptReader& reader, ScriptBookmark& bm, TapeWriter& writer);
+   void parseBuildScriptArgumentList(_ScriptReader& reader, Stack<ScriptBookmark>& callStack);
+   void parseBuildScriptStatement(_ScriptReader& reader, Stack<ScriptBookmark>& callStack);
 
-   bool parse(TapeWriter& writer, _ScriptReader& reader);
-   void parsePostfix(TapeWriter& writer);
+   void parseInlineStatement(_ScriptReader& reader, ScriptBookmark& bm, TapeWriter& writer);
+
+   void parseInlineScript(TapeWriter& writer, _ScriptReader& reader);
+   bool parseBuildScript(TapeWriter& writer, _ScriptReader& reader);
+   void parseInline(ident_t script, TapeWriter& writer);
 
 public:
    virtual bool setPostfix(ident_t postfix_script)