+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - ELC
   - a warning shown when an object cannot be converted into another one
   - a warning shown when an object does not handle a message (for strong-typed ones)
+- TOOLS
+   - #323 : Run ELENA code without build
 
 ## ELENA 5.0
 *14.07.2020*